0.1.6: hotfixes and some small missing features.
This is because we need to (internally) allocate an array "Mem" of values corresponding to nodes.
Previously `eval` would simply fail for `0 → n` maps.

A better approach would be to only keep in memory those nodes which will be needed later.
However, this requires more information from the layout algorithm. TODO!

Bug: the wrong offset was added to each node ID in quotient map.

Same as strict version; here for convenience.